Fix body check in rest.py
Checking for "trueness" here is not sufficient.  An empty list has a truth
value of false so it will not be encoded as json and sent as the request
body as it should.  Instead we need to check explicitly if body is None.

For example, an empty list is sent when deleting all presubscriptions
via update_presubscriptions.